_showing data from local database; _persisting only projects (+tree), issues (+tree), time entries (+tree); _built base for on demand single entity (+tree) refresh;
ItemViewProperties.java: small changes;
NetworkingService.java: prepared for on demand entity refresh;
Repository.java: added type definition to class name so that implementations will have by default the needed return types; moved generic (reusable) code from concrete classes;
ResizeableItemsUiOps.java: added 'lifecycle' events for recycler view width changing;
TimeEntriesAdapter.java: show local db data; updates when changes are detected in the TimeEntries db table;
TimeEntriesRepository.java: added logic to register to db changes and to network service periodic updates (disabled for now); only inserts/updates data for now;
TimeEntriesUiOps.java: now doesn't have an adapter when its not showing; linked with Repository, starts and stops it according to parent state changes;
IssuesViewHolder.java: added entity id;
IssueData.java: changed id field from long to int;
IssuesRepository.java: added logic to register to db changes and to network service periodic updates (disabled for now); only inserts/updates data for now;
IssuesUiOps.java: now doesn't have an adapter when its not showing; linked with Repository, starts and stops it according to parent state changes;
TimeEntryStorIOSQLiteGetResolver.java: now gets the linked user data too;
ProjectsViewHolder.java: added entity id;
ProjectData.java: changed id field from long to int;
ProjectsRepository.java: moved generic methods to the abstract class; removes everything + inserts all when a periodic update happens;
BaseInterfaces.java: added RvResizeListener with two methods that are (supposed to be) called when someone wants to be announced of a rv width change;
activity_home_scrolling.xml: removed unused commented code;

+++ app/src/main/java/org/rares/miner49er/issues/model/IssueData.java
@@ -13,8 +13,8 @@ import java.util.List;
 @Data
 public class IssueData {
 
-    private long id;
-    private long projectId;
+    private int id;
+    private int projectId;
     private String name;
     //    private User owner;
     private long dateAdded;
